Move newly added components out of Version2 due to the failing JS test.
