+++ assets/js/modules/analytics/components/dashboard/DimensionTabs.js
